chore(dashboard): remove unused assets and refactor imports
- Deleted placeholder logo images (PNG, SVG, JPG) from the public directory.
- Refactored import paths in various action files and components to streamline access to shared types and constants.
- Removed obsolete utility functions and types related to toast notifications and layout utilities.

This cleanup enhances project organization and reduces unnecessary file clutter.
