+++ dashboard/src/app/[tenantId]/ExternalEventDefs/[externalEventDefId]/page.tsx
@@ -1,4 +1,4 @@
-import { lhClient } from '@/utils/client/lhClient'
+import { lhClient } from '@/lhClient'
 import ExternalEventDefClient from './ExternalEventDefClient'
 
 interface ExternalEventDefPageProps {
