+++ dashboard/src/app/[tenantId]/UserTaskDefs/[userTaskDefId]/UserTaskDefClient.tsx
@@ -3,7 +3,7 @@
 import { searchUserTaskRun } from '@/actions/searchUserTaskRun'
 import LinkWithTenant from '@/components/link-with-tenant'
 import { Pagination } from '@/components/ui/load-more-pagination'
-import { SEARCH_LIMIT_DEFAULT, SEARCH_LIMITS } from '@/utils/ui/constants'
+import { SEARCH_LIMIT_DEFAULT, SEARCH_LIMITS } from '@/constants'
 import { Badge } from '@littlehorse-enterprises/ui-library/badge'
 import { Card, CardContent, CardHeader, CardTitle } from '@littlehorse-enterprises/ui-library/card'
 import {
