+++ dashboard/src/components/diagram/left-sidebar/tab-content/wf-run-tab.tsx
@@ -5,7 +5,7 @@ import { DataTable } from '@/components/ui/data-table'
 import { DropdownFilter } from '@/components/ui/dropdown-filter'
 import { FilterResetButton } from '@/components/ui/filter-reset-button'
 import { useExecuteRPCWithSWR } from '@/hooks/useExecuteRPCWithSWR'
-import { MINUTES_TO_TIME_RANGE, STATUS_OPTIONS, TIME_RANGE_MINUTES, TIME_RANGE_OPTIONS } from '@/utils/ui/constants'
+import { MINUTES_TO_TIME_RANGE, STATUS_OPTIONS, TIME_RANGE_MINUTES, TIME_RANGE_OPTIONS } from '@/constants'
 import type { ColumnDef } from '@tanstack/react-table'
 import { LHStatus, WfRun } from 'littlehorse-client/proto'
 import { CheckCircle, Clock, Loader2, XCircle } from 'lucide-react'
