+++ dashboard/src/components/diagram/left-sidebar/tab-content/wf-spec-tab.tsx
@@ -2,9 +2,9 @@
 
 import { useState } from 'react'
 import { WfSpec, WfRun } from 'littlehorse-client/proto'
-import { buildNodeTree, sortNodeTree } from '@/utils/data/node-tree'
 import { TreeNodeComponent } from '../../tree-node-component'
 import { NodeSearchBar } from '../../node-search-bar'
+import { TreeNode } from '@/types'
 
 interface WfSpecTabProps {
   wfSpec: WfSpec
@@ -32,4 +32,36 @@ export default function WfSpecTab({ wfSpec }: WfSpecTabProps) {
       </div>
     </div>
   )
+}
+
+function buildNodeTree(wfSpec: WfSpec, threadSpecName: string): TreeNode[] {
+  const result: TreeNode[] = []
+
+  Object.entries(wfSpec.threadSpecs[threadSpecName].nodes).forEach(([nodeId]) => {
+    result.push({
+      id: `${nodeId}:${threadSpecName}`,
+      label: nodeId,
+      type: undefined,
+      status: undefined,
+      children: [],
+      level: 0,
+    })
+  })
+
+  return result
+}
+
+function sortNodeTree(nodeTree: TreeNode[]): TreeNode[] {
+  return [...nodeTree].sort((a, b) => {
+    const aFirstPart = a.id.split('-')[0]
+    const bFirstPart = b.id.split('-')[0]
+
+    // If the first part is a number, sort numerically in ascending order
+    if (!isNaN(Number(aFirstPart)) && !isNaN(Number(bFirstPart))) {
+      return Number(aFirstPart) - Number(bFirstPart)
+    }
+
+    // Fallback to alphabetical sorting if not numbers
+    return bFirstPart.localeCompare(aFirstPart)
+  })
 }
