+++ dashboard/src/components/metadata/metadata-search-client.tsx
@@ -2,8 +2,8 @@
 import { executeRpc } from '@/actions/executeRPC'
 import { search, SearchResponse } from '@/actions/search'
 import { Pagination } from '@/components/ui/load-more-pagination'
-import { SearchType } from '@/types/search'
-import { SEARCH_ENTITIES, SEARCH_LIMIT_DEFAULT, SEARCH_LIMITS } from '@/utils/ui/constants'
+import { SearchType } from '@/types'
+import { SEARCH_ENTITIES, SEARCH_LIMIT_DEFAULT, SEARCH_LIMITS } from '@/constants'
 import { Tabs, TabsList, TabsTrigger } from '@littlehorse-enterprises/ui-library/tabs'
 import { WfSpecId } from 'littlehorse-client/proto'
 import { useParams } from 'next/navigation'
