--- dashboard/src/types/PageParams.ts
@@ -1,6 +1,0 @@
-export type PageParams = {
-  params: Promise<unknown>
-  searchParams: Promise<unknown>
-}
-
-export type PathnameKeys = keyof Awaited<PageParams['params']>
