--- dashboard/src/types/leftSidebarTabs.ts
@@ -1,1 +1,0 @@
-export type LeftSidebarTabId = 'WfSpec' | 'WfRuns' | 'ScheduledWfRuns'
