--- dashboard/src/types/search.ts
@@ -1,2 +1,0 @@
-import { SEARCH_ENTITIES } from '@/utils/ui/constants'
-export type SearchType = (typeof SEARCH_ENTITIES)[number]
