--- dashboard/src/types/withs.ts
@@ -1,6 +1,0 @@
-export type WithTenant = {
-  tenantId: string
-}
-export type WithBookmark = {
-  bookmark?: string
-}
