--- dashboard/src/utils/ui/layout-utils.ts
@@ -1,31 +1,0 @@
-import dagre from '@dagrejs/dagre'
-import { CustomNode, CustomEdge } from '@/types/node'
-
-const dagreGraph = new dagre.graphlib.Graph()
-
-export function getLayoutedElements(nodes: CustomNode[], edges: CustomEdge[]) {
-  // #region GraphSetup
-  dagreGraph.setGraph({ rankdir: 'LR', align: 'UR' })
-
-  nodes.forEach(node => {
-    dagreGraph.setNode(node.id, { width: 100, height: 100 })
-  })
-  edges.forEach(edge => {
-    dagreGraph.setEdge(edge.source, edge.target, { width: edge.label ? 200 : 10 })
-  })
-
-  dagre.layout(dagreGraph)
-  // #endregion
-
-  const layoutedNodes = nodes.map(node => {
-    const dagreNode = dagreGraph.node(node.id)
-
-    return {
-      ...node,
-      position: { x: dagreNode.x - 100 / 2, y: dagreNode.y - 100 / 2 },
-      layouted: true,
-    }
-  })
-
-  return { nodes: layoutedNodes, edges }
-}
